# api/openapi.yaml
openapi: 3.0.3
info:
  title: IGNAI Hackathon Platform API
  version: 0.4.0
  description: 'Spring Boot / MyBatis 版本接口规划。MVP 无独立评委角色，无视频上传； 后台管理入口规划为 /root/，root
    初始化和认证使用 /api/root/*。

    '
servers:
- url: http://localhost:8080/api
  description: Local Spring Boot API
...
paths:
...
  /public/events/{eventId}/form-fields:
    get:
      tags:
      - Public
      summary: 获取动态表单字段
      description: |
        ## 获取活动的自定义表单字段
        返回指定活动下，报名表或作品提交表的字段配置。前端根据返回的字段列表动态渲染表单。
        - **无需认证**：公开接口，任何人可访问
        - **target=registration**：返回报名表单字段（如姓名、手机号、GitHub 链接等）
        - **target=project**：返回作品提交表单字段（如项目名、演示链接、代码仓库等）
        - **排序**：字段按 `sortOrder` 升序排列，只返回 `enabled=1` 的字段
      parameters:
      - $ref: '#/components/parameters/EventId'
      - name: target
        in: query
        required: true
        schema:
          type: string
          enum:
          - registration
          - project
      responses:
        '200':
          description: 字段列表
          content:
            application/json:
              schema:
                type: object
      operationId: public_events_eventId_form-fields_get
...
  /public/projects:
    get:
      tags:
      - Project
      summary: 公开作品列表
      description: |
        ## 前台作品展示列表
        返回已发布或已提交的作品，支持多维度筛选和排序。
        - **无需认证**：访客浏览作品墙
        - **筛选**：`status`（published/submitted）、`trackId`（赛道）、`sort`（latest/votes/rating/featured）
        - **用途**：前台作品展示页、排行榜
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - published
          - submitted
      - name: trackId
        in: query
        schema:
          type: integer
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - latest
          - votes
          - rating
          - featured
      - name: pageSize
        in: query
        schema:
          type: integer
      responses:
        '200':
          description: 作品列表
          content:
            application/json:
              schema:
                type: object
      operationId: public_projects_get
  /public/projects/{projectId}:
    get:
      tags:
      - Project
      summary: 公开作品详情
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 作品详情
          content:
            application/json:
              schema:
                type: object
      operationId: public_projects_projectId_get
  /public/projects/{projectId}/votes:
    post:
      tags:
      - Rating
      summary: 登录用户点赞作品
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 点赞成功
          content:
            application/json:
              schema:
                type: object
        '409':
          description: 重复点赞
          content:
            application/json:
              schema:
                type: object
      operationId: public_projects_projectId_votes_post
  /public/projects/{projectId}/ratings:
    post:
      tags:
      - Rating
      summary: 开放评分
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingCreateRequest'
      responses:
        '200':
          description: 评分成功
          content:
            application/json:
              schema:
                type: object
        '409':
          description: 重复评分
          content:
            application/json:
              schema:
                type: object
      operationId: public_projects_projectId_ratings_post
...
  /projects:
    post:
      tags:
      - Project
      summary: 创建或保存作品
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpsertRequest'
      responses:
        '200':
          description: 作品信息
          content:
            application/json:
              schema:
                type: object
      operationId: projects_post
...
  /projects/{projectId}:
    put:
      tags:
      - Project
      summary: 更新作品
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
      operationId: projects_projectId_put
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: integer
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ProjectUpsertRequest:
      type: object
      required:
      - eventId
      - teamId
      - title
      properties:
        eventId:
          type: integer
        teamId:
          type: integer
        trackId:
          type: integer
        title:
          type: string
        tagline:
          type: string
        description:
          type: string
        demoUrl:
          type: string
        codeUrl:
          type: string
        payload:
          type: object
...
    RatingCreateRequest:
      type: object
      required:
      - score
      - raterKey
      properties:
        score:
          type: number
          minimum: 1
          maximum: 5
        raterKey:
          type: string
        sourceType:
          type: string
          enum:
          - public
          - participant
          - onsite_code
...